refactor(compiler): move paramShape metadata into semantic
