+++ packages/compiler/src/semantic/buildNamespace.ts
@@ -22,7 +22,6 @@ import {
 	type ValidatedPrototypeAST,
 } from '@8f4e/compiler-spec';
 import { getError } from '../compilerError';
-import { getEffectiveFunctionMetadata } from '../prototypes/paramShape';
 import { createCompilationContext } from './createCompilationContext';
 import { applyMemoryDeclarationLine } from './declarations';
 import applySemanticInstruction from './instructions';
@@ -35,6 +34,7 @@ import {
 	validateMemoryRegionOptions,
 } from './memoryRegions';
 import normalizeCompileTimeArguments from './normalizeCompileTimeArguments';
+import { getEffectiveFunctionMetadata } from './paramShape';
 import parseMemoryInstructionArguments from './utils/memoryInstructionParser';
 
 const moduleBlock = compilerSourceBlockInstructionByType.module;
